requirements.txt
tensorflow-cpu==2.6.0
keras==2.6.0
numpy==1.19.5
protobuf==3.20.3
flask==2.0.3
pillow==8.3.1
h5py==3.1.0
